+++ problema02/src/main/java/br/furb/problema02/order/OrderTypeSell.java
@@ -3,8 +3,10 @@ package br.furb.problema02.order;
 import br.furb.problema02.enums.OrderTypeEnum;
 import br.furb.problema02.model.Order;
 
+import java.math.BigDecimal;
+
 public class OrderTypeSell extends Order{
-	public OrderTypeSell(String investorName, double orderValue) {
+	public OrderTypeSell(String investorName, BigDecimal orderValue) {
 		super(investorName, orderValue);
 	}
